+++ applets/0CinnamonMenuForkByOdyseus/__data__/CHANGELOG.md
@@ -4,6 +4,19 @@
 
 ***
 
+- **Date:** Tue, 12 Feb 2019 23:52:51 -0300
+- **Commit:** [3ed94a6](https://gitlab.com/Odyseus/CinnamonTools/commit/3ed94a6)
+- **Author:** Odyseus
+
+```
+- Removed some try{}catch{} blocks left for debugging.
+- Changed some defaults to suit my needs.
+- Several minor optimizations and clean ups (most of them from upstream).
+
+```
+
+***
+
 - **Date:** Mon, 21 Jan 2019 21:50:38 -0300
 - **Commit:** [ebb12e6](https://gitlab.com/Odyseus/CinnamonTools/commit/ebb12e6)
 - **Author:** Odyseus